refactor(frontend): align benchmark dataset gallery
